chore(socket): moved around test files
