+++ src/socket/.gitignore
@@ -1,0 +1,2 @@
+coverage
+.nyc_output
