+++ src/socket/package.json
@@ -4,30 +4,30 @@
   "description": "",
   "main": "ecosystem.config.js",
   "scripts": {
-    "test": "NODE_ENV=testing && node_modules/.bin/mocha -r ts-node/register --exit test/**/*.ts",
+    "test": "NODE_ENV=testing && node_modules/.bin/mocha -r ts-node/register --exit test/**/*_test.ts",
     "start": "pm2-runtime ecosystem.config.js",
     "cov": "node_modules/.bin/nyc npm run test",
     "restart": "pm2 restart all",
     "build": "npm ci && node_modules/.bin/tsc"
   },
   "dependencies": {
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "socket.io": "^2.3.0"
+  },
+  "devDependencies": {
+    "ts-node": "^8.8.1",
+    "typescript": "^3.8.3",
+    "nyc": "^15.0.1",
+    "mocha": "^7.1.1",
     "@types/express": "^4.17.3",
     "@types/node": "^13.9.0",
     "@types/socket.io": "^2.1.4",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "mocha": "^7.1.1",
-    "nyc": "^15.0.1",
-    "socket.io": "^2.3.0",
-    "ts-node": "^8.8.1",
-    "typescript": "^3.8.3"
-  },
-  "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "0.1.3",
     "@types/chai": "^4.2.11",
     "@types/mocha": "^7.0.2",
     "@types/socket.io-client": "^1.4.32",
-    "chai": "^4.2.0",
+    "chai": "^4.2.0"
   },
   "author": "",
   "license": "ISC"
