+++ src/socket/tsconfig.json
@@ -61,6 +61,12 @@
     // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
 
     /* Advanced Options */
-    "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */
-  }
+    "forceConsistentCasingInFileNames": true,  /* Disallow inconsistently-cased references to the same file. */
+  },
+  "include": [
+    "server.ts",
+    "app.ts",
+    "socket.ts",
+    "bin/www.ts"
+  ]
 }
